refactor: Move lora nodes to individual files
Implements ML-597
